worker: move job#Spawn => worker.Spawn()
