+++ jobqueue/jobqueue.go
@@ -3,6 +3,7 @@ package jobqueue
 import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/codestand/build/job"
+	"github.com/codestand/build/worker"
 )
 
 var queue chan job.Job
@@ -27,7 +28,7 @@ func Push(j job.Job) {
 func Wait() {
 	for {
 		if j, ok := <-queue; ok {
-			if err := j.Spawn(); err != nil { // TODO: parallelize
+			if err := worker.Spawn(j); err != nil { // TODO: parallelize
 				log.Fatal(err) // TODO: improve error handling
 			}
 		} else {
